requirements.txt
anykeystore==0.2
apex==0.1
aspy.refactor-imports==2.1.1
asv==0.4.1
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
configparser==5.0.0
coolname==1.1.0
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.16.0
data==0.4
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
dominate==2.5.1
easydict==1.9
funcsigs==1.0.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
gql==0.2.0
graphql-core==1.1
hupper==1.10.2
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
jsonpatch==1.25
jsonpointer==2.0
kdtree==0.16
kiwisolver==1.2.0
latex==0.7.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.3
PasteDeploy==2.1.0
pathtools==0.1.2
pbkdf2==1.3
Pillow==7.1.2
plaster==1.0
plaster-pastedeploy==0.7
promise==2.3
psutil==5.7.0
pycocotools==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pyellipsoid==0.0.2
pyflakes==2.2.0
pyparsing==2.4.7
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
reorder-python-imports==2.3.1
repoze.sendmail==4.4.1
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scipy==1.4.1
sentry-sdk==0.14.4
shortuuid==1.0.1
shutilwhich==1.1.0
sip==4.19.13
six==1.14.0
smmap==3.0.4
SQLAlchemy==1.3.16
subprocess32==3.5.4
tempdir==0.7.1
toolz==0.10.0
torch==1.6.0a0+e3fcc6a
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.46.0
transaction==3.0.0
translationstring==1.3
urllib3==1.25.9
velruse==1.1.1
venusian==3.0.0
visdom==0.1.8.9
wandb==0.8.36
watchdog==0.10.2
WebOb==1.8.6
websocket-client==0.57.0
WTForms==2.3.1
wtforms-recaptcha==0.3.2
yapf==0.30.0
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
